+++ libraries/FunctionGenerator/FunctionGenerator.h
@@ -2,14 +2,14 @@
 //
 //    FILE: functionGenerator.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.6
+// VERSION: 0.3.0
 // PURPOSE: wave form generating functions (use with care)
 //     URL: https://github.com/RobTillaart/FunctionGenerator
 
 
 #include "Arduino.h"
 
-#define FUNCTIONGENERATOR_LIB_VERSION           (F("0.2.6"))
+#define FUNCTIONGENERATOR_LIB_VERSION           (F("0.3.0"))
 
 
 class funcgen
@@ -53,7 +53,7 @@ public:
   float zero();
 
   //  standard wave forms
-  float sawtooth(float t, uint8_t mode = 0);   //  0 ==>  /|.   1 ==> sawtooth |\.
+  float sawtooth(float t, uint8_t mode = 0);  //  0 ==>  /|.   1 ==> sawtooth |\.
   float triangle(float t);
   float square(float t);
   float sinus(float t);
@@ -61,6 +61,17 @@ public:
 
   float random();
   float random_DC();  //  duty cycle variant. Experimental.
+
+  /////////////////////////////////////////////////////////////
+  //
+  //  EXPERIMENTAL 0.2.7
+  //
+  float sinusDiode(float t);
+  float sinusRectified(float t);
+  float trapezium1(float t);
+  float trapezium2(float t);
+  float heartBeat(float t);  //  72 BPM = 72/60 = 1 setFrequency(1.2)
+  float freeWave(float t, int16_t * arr, int16_t size);  //  arr must be size+1 long
 
 
 private: